Refactor meal planner action hooks
